added more logging

+++ lib/screens/manager/parties/party_add_edit_screen.dart
@@ -367,7 +367,9 @@ class _PartyAddEditScreenState extends State<PartyAddEditScreen> {
                 onClicked: () {
                   if (isPhotoChanged) {
                     widget.party = widget.party.copyWith(imageUrl: newImageUrl);
-                    FirestorageHelper.deleteFile(oldImageUrl);
+                    if(oldImageUrl.isNotEmpty) {
+                      FirestorageHelper.deleteFile(oldImageUrl);
+                    }
                   }
 
                   if (widget.party.blocServiceId.isEmpty) {
